Switch bsp loading to engine-agnostic crate
