+++ src/prelude.rs
@@ -15,8 +15,6 @@ pub(crate) use std::mem;
 pub(crate) use std::path::{Path, PathBuf};
 pub(crate) use std::sync::*;
 pub(crate) use thiserror::Error;
-pub(crate) use rayon::prelude::*;
-pub(crate) use grid::*;
 
 pub use anyhow;
 pub use indexmap;
